// src/exchanges/public-com/client/types.ts
export enum InstrumentType {
  ALT = 'ALT',
  BOND = 'BOND',
  CRYPTO = 'CRYPTO',
  EQUITY = 'EQUITY',
  INDEX = 'INDEX',
  MULTI_LEG_INSTRUMENT = 'MULTI_LEG_INSTRUMENT',
  OPTION = 'OPTION',
  TREASURY = 'TREASURY',
}
// ...
export enum OrderSide {
  BUY = 'BUY',
  SELL = 'SELL',
}

export enum OrderType {
  MARKET = 'MARKET',
  LIMIT = 'LIMIT',
  STOP = 'STOP',
  STOP_LIMIT = 'STOP_LIMIT',
}

export enum TimeInForce {
  DAY = 'DAY',
  GTD = 'GTD',
}
// ...
export enum OrderStatus {
  NEW = 'NEW',
  PARTIALLY_FILLED = 'PARTIALLY_FILLED',
  CANCELLED = 'CANCELLED',
  QUEUED_CANCELLED = 'QUEUED_CANCELLED',
  FILLED = 'FILLED',
  REJECTED = 'REJECTED',
  PENDING_REPLACE = 'PENDING_REPLACE',
  PENDING_CANCEL = 'PENDING_CANCEL',
  EXPIRED = 'EXPIRED',
  REPLACED = 'REPLACED',
}